new start command
